requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
asn1crypto==0.24.0
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
catfish==1.4.4
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==3.3.1
cupshelpers==1.0
daphne==3.0.1
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
filelock==3.0.12
httplib2==0.9.2
hyperlink==20.0.1
idna==3.1
importlib-metadata==3.1.1
importlib-resources==3.3.0
incremental==17.5.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
menulibre==2.2.0
mugshot==0.4.0
netifaces==0.10.4
numpy==1.19.4
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
pexpect==4.2.1
Pillow==5.1.0
pipenv==2020.11.15
psutil==5.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygame==2.0.0
pygobject==3.26.1
PyHamcrest==2.0.2
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pyOpenSSL==20.0.1
python-apt==1.6.5+ubuntu0.4
python-dateutil==2.6.1
python-debian==0.1.32
python-debianbts==2.7.2
python-dotenv==0.15.0
python-magic==0.4.16
pytz==2020.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
sgt-launcher==0.2.4
simplejson==3.13.2
six==1.15.0
sqlparse==0.4.1
systemd-python==234
Twisted==20.3.0
txaio==20.12.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==20.2.2
virtualenv-clone==0.5.4
wadllib==1.3.2
xcffib==0.5.1
xkit==0.0.0
zipp==3.4.0
zope.interface==5.2.0
